pyproject.toml
[project]
name = "django-next-baseline"
version = "0.1.0"
description = "A Django backend-first baseline prepared for a Next.js frontend."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2,<5.3",
    "django-environ>=0.12,<0.13",
    "django-cors-headers>=4.9,<5",
    "django-filter>=25.2,<26",
    "drf-spectacular>=0.28,<0.29",
    "djangorestframework>=3.16,<3.17",
    "djangorestframework-simplejwt>=5.5,<6",
    "openpyxl>=3.1,<3.2",
    "psycopg[binary]>=3.2,<3.3",
    "whitenoise>=6.9,<7",
]

[dependency-groups]
dev = [
    "django-extensions>=4.1,<5",
    "mypy>=1.17,<1.18",
    "pre-commit>=4.3,<5",
    "pytest>=9.0,<10",
    "pytest-django>=4.11,<5",
    "ruff>=0.12,<0.13",
    "uv>=0.11,<0.12",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "tests.py"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"
exclude = [".venv", "staticfiles", "media", "logs", "apps/system/migrations"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
exclude = ["apps/system/migrations", ".venv", "staticfiles"]
check_untyped_defs = false
files = ["apps/common", "apps/accounts", "apps/audit", "apps/approval", "config"]
